오류 수정
Stat Upgrade의 stat 정보가 제대로 저장되지 않는 오류 수정
UI 효과음 구현 중

+++ Roguelike-game-v2/Assets/Scripts/Etc/StatSelection.cs
@@ -3,17 +3,17 @@ public class StatSelection : MonoBehaviour
 {
     public const int maxLevel = 5;
 
-    private int moveSpeed = 0;
-    private int increaseHealth = 0;
-    private int increaseDamage = 0;
-    private int healthRegenPerSec = 0;
+    private float moveSpeed = 0;
+    private float increaseHealth = 0;
+    private float increaseDamage = 0;
+    private float healthRegenPerSec = 0;
 
     public void Set(PlayerStat stat)
     {
-        moveSpeed = (int)stat.MoveSpeed;
-        increaseHealth = (int)stat.IncreaseHealth;
-        increaseDamage = (int)stat.IncreaseDamage;
-        healthRegenPerSec = (int)stat.HealthRegenPerSec;
+        moveSpeed = stat.MoveSpeed;
+        increaseHealth = stat.IncreaseHealth;
+        increaseDamage = stat.IncreaseDamage;
+        healthRegenPerSec = stat.HealthRegenPerSec;
     }
     public void Save()
     {
